+++ astra/linux/libbcm_tzioc/libtzioc.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- * Broadcom Proprietary and Confidential. (c)2016 Broadcom. All rights reserved.
+ * Copyright (C) 2017 Broadcom.  The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
  *
  * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
@@ -129,6 +129,6 @@ uint32_t _tzioc_vaddr2offset(
 
 int _tzioc_call_smc(
     struct tzioc_client *pClient,
-    uint8_t ucMode);
+    uint32_t ucMode);
 
 #endif /* LIBTZIOC_H */
