+++ astra/linux/libbcm_tzioc/libtzioc_ioctl.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- * Broadcom Proprietary and Confidential. (c)2016 Broadcom. All rights reserved.
+ * Copyright (C) 2017 Broadcom.  The term "Broadcom" refers to Broadcom Limited and/or its subsidiaries.
  *
  * This program is the proprietary software of Broadcom and/or its licensors,
  * and may only be used, duplicated, modified or distributed pursuant to the terms and
@@ -63,6 +63,6 @@ int _tzioc_ioctl_mem_free(
 
 int _tzioc_ioctl_call_smc(
     struct tzioc_client *pClient,
-    uint8_t ucMode);
+    uint32_t ucMode);
 
 #endif /* LIBTZIOC_H */
